+++ src/main/java/com/github/hexirp/forge/prp/core/recipes/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * PRP-Coreで追加されるレシピを定義するパッケージ.
+ *
+ * @author Hexirp
+ */
+package com.github.hexirp.forge.prp.core.recipes;
